Update CRD to include init job options
