+++ crds/vshn.appcat.vshn.io_xvshnopenbaos.yaml
@@ -5418,11 +5418,27 @@ spec:
                         description: Init contains settings to control the init job
                           of this OpenBao instance.
                         properties:
+                          backoffLimit:
+                            default: 3
+                            description: |-
+                              BackoffLimit defines how many times the init job may be retried before
+                              it is considered failed. Defaults to 3.
+                            format: int32
+                            minimum: 0
+                            type: integer
                           enabled:
                             default: true
                             description: Enabled enables or disables the automatic
                               init job for this OpenBao instance.
                             type: boolean
+                          ttlSecondsAfterFinished:
+                            default: 86400
+                            description: |-
+                              TTLSecondsAfterFinished defines how many seconds the init job and its pod
+                              are kept after the job finishes. Defaults to 86400 (1 day).
+                            format: int32
+                            minimum: 0
+                            type: integer
                         type: object
                       serviceLevel:
                         default: besteffort
@@ -6019,6 +6035,8 @@ spec:
                       type: string
                   type: object
                 type: array
+            required:
+            - initialized
             type: object
         required:
         - spec
