+++ pkg/comp-functions/functions/vshnopenbao/deploy.go
@@ -9,20 +9,16 @@ import (
 	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 	xfnproto "github.com/crossplane/function-sdk-go/proto/v1"
 	xhelmbeta1 "github.com/vshn/appcat/v4/apis/helm/release/v1beta1"
-	xkube "github.com/vshn/appcat/v4/apis/kubernetes/v1alpha2"
 	vshnv1 "github.com/vshn/appcat/v4/apis/vshn/v1"
 	"github.com/vshn/appcat/v4/pkg/common/utils"
 	"github.com/vshn/appcat/v4/pkg/comp-functions/runtime"
-	corev1 "k8s.io/api/core/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8sruntime "k8s.io/apimachinery/pkg/runtime"
 )
 
-const (
-	initOutputSecretSuffix = "-init-output"
-	unsealKeysSecretSuffix = "-unseal-keys"
-	observerSuffix         = "-observer"
-)
+const serverRoleSuffix = "-server"
+
 
 func DeployOpenBao(ctx context.Context, comp *vshnv1.VSHNOpenBao, svc *runtime.ServiceRuntime) *xfnproto.Result {
 	if err := svc.GetObservedComposite(comp); err != nil {
@@ -178,88 +174,53 @@ func DeployOpenBao(ctx context.Context, comp *vshnv1.VSHNOpenBao, svc *runtime.S
 		return runtime.NewWarningResult(fmt.Errorf("cannot add %s composed resource: %w", comp.Name+"-release", err).Error())
 	}
 
-	if err := observeInitConnectionDetails(comp, svc); err != nil {
-		return runtime.NewFatalResult(fmt.Errorf("cannot observe init connection details: %w", err))
+	if err := configureDiscoveryRBAC(serviceName, comp.GetInstanceNamespace(), svc); err != nil {
+		return runtime.NewWarningResult(fmt.Errorf("cannot configure discovery RBAC: %w", err).Error())
 	}
 
 	return nil
 }
 
-func observeInitConnectionDetails(comp *vshnv1.VSHNOpenBao, svc *runtime.ServiceRuntime) error {
-	serviceName := comp.GetName()
-	ns := comp.GetInstanceNamespace()
+func configureDiscoveryRBAC(serviceName, ns string, svc *runtime.ServiceRuntime) error {
+	roleName := serviceName + serverRoleSuffix
 
-	rootTokenSecretName := serviceName + initOutputSecretSuffix
-	unsealKeysSecretName := serviceName + unsealKeysSecretSuffix
-
-	rootTokenSecret := &corev1.Secret{
+	role := &rbacv1.Role{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      rootTokenSecretName,
+			Name:      roleName,
 			Namespace: ns,
 		},
-	}
-	err := svc.SetDesiredKubeObject(rootTokenSecret, rootTokenSecretName+observerSuffix,
-		runtime.KubeOptionObserve,
-		runtime.KubeOptionAllowDeletion,
-		runtime.KubeOptionAddConnectionDetails(svc.GetCrossplaneNamespace(),
-			xkube.ConnectionDetail{
-				ObjectReference: corev1.ObjectReference{
-					APIVersion: "v1",
-					Kind:       "Secret",
-					Namespace:  ns,
-					Name:       rootTokenSecretName,
-					FieldPath:  "data.VAULT_ADDR",
-				},
-				ToConnectionSecretKey: "VAULT_ADDR",
+		Rules: []rbacv1.PolicyRule{
+			{
+				APIGroups: []string{""},
+				Resources: []string{"pods"},
+				Verbs:     []string{"get", "watch", "list", "update", "patch"},
 			},
-			xkube.ConnectionDetail{
-				ObjectReference: corev1.ObjectReference{
-					APIVersion: "v1",
-					Kind:       "Secret",
-					Namespace:  ns,
-					Name:       rootTokenSecretName,
-					FieldPath:  "data.VAULT_TOKEN",
-				},
-				ToConnectionSecretKey: "VAULT_TOKEN",
-			},
-		),
-	)
-	if err != nil {
-		return fmt.Errorf("cannot set root token secret observer: %w", err)
-	}
-
-	if err = svc.AddObservedConnectionDetails(rootTokenSecretName + observerSuffix); err != nil {
-		return fmt.Errorf("cannot add root token connection details: %w", err)
-	}
-
-	unsealKeysSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      unsealKeysSecretName,
-			Namespace: ns,
 		},
 	}
-	err = svc.SetDesiredKubeObject(unsealKeysSecret, unsealKeysSecretName+observerSuffix,
-		runtime.KubeOptionObserve,
-		runtime.KubeOptionAllowDeletion,
-		runtime.KubeOptionAddConnectionDetails(svc.GetCrossplaneNamespace(),
-			xkube.ConnectionDetail{
-				ObjectReference: corev1.ObjectReference{
-					APIVersion: "v1",
-					Kind:       "Secret",
-					Namespace:  ns,
-					Name:       unsealKeysSecretName,
-					FieldPath:  "data.keys",
-				},
-				ToConnectionSecretKey: "keys",
-			},
-		),
-	)
-	if err != nil {
-		return fmt.Errorf("cannot set unseal keys secret observer: %w", err)
+	if err := svc.SetDesiredKubeObject(role, roleName); err != nil {
+		return fmt.Errorf("cannot add discovery role: %w", err)
 	}
 
-	if err = svc.AddObservedConnectionDetails(unsealKeysSecretName + observerSuffix); err != nil {
-		return fmt.Errorf("cannot add unseal keys connection details: %w", err)
+	rb := &rbacv1.RoleBinding{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      roleName,
+			Namespace: ns,
+		},
+		Subjects: []rbacv1.Subject{
+			{
+				Kind:      "ServiceAccount",
+				Name:      serviceName,
+				Namespace: ns,
+			},
+		},
+		RoleRef: rbacv1.RoleRef{
+			APIGroup: "rbac.authorization.k8s.io",
+			Kind:     "Role",
+			Name:     roleName,
+		},
+	}
+	if err := svc.SetDesiredKubeObject(rb, roleName+"-binding"); err != nil {
+		return fmt.Errorf("cannot add discovery rolebinding: %w", err)
 	}
 
 	return nil
