+++ pkg/comp-functions/functions/vshnopenbao/utils.go
@@ -28,3 +28,23 @@ func getOpenBaoTestCompWithInitSecret(t *testing.T) (*runtime.ServiceRuntime, *v
 
 	return svc, comp
 }
+
+func getOpenBaoTestCompWithInitDisabled(t *testing.T) (*runtime.ServiceRuntime, *vshnv1.VSHNOpenBao) {
+	svc := commontest.LoadRuntimeFromFile(t, "vshnopenbao/deploy/03_init_disabled.yaml")
+
+	comp := &vshnv1.VSHNOpenBao{}
+	err := svc.GetObservedComposite(comp)
+	assert.NoError(t, err)
+
+	return svc, comp
+}
+
+func getOpenBaoTestCompWithStatusInitialized(t *testing.T) (*runtime.ServiceRuntime, *vshnv1.VSHNOpenBao) {
+	svc := commontest.LoadRuntimeFromFile(t, "vshnopenbao/deploy/04_status_initialized.yaml")
+
+	comp := &vshnv1.VSHNOpenBao{}
+	err := svc.GetObservedComposite(comp)
+	assert.NoError(t, err)
+
+	return svc, comp
+}
