+++ test/functions/vshnopenbao/deploy/03_init_disabled.yaml
@@ -29,35 +29,52 @@ observed:
             version: "2.0"
             serviceLevel: besteffort
             init:
-              enabled: true
+              enabled: false
           tls:
             enabled: true
             authClients: true
       status:
         instanceNamespace: vshn-openbao-openbao-test
   resources:
-    openbao-test-init-output-observer:
+    openbao-test-claim-ns-observer:
       resource:
         apiVersion: kubernetes.crossplane.io/v1alpha2
         kind: Object
         metadata:
-          name: openbao-test-init-output
+          name: unit-test
         spec:
           forProvider:
             manifest:
               apiVersion: v1
-              kind: Secret
+              kind: Namespace
               metadata:
-                name: openbao-test-init-output
-                namespace: vshn-openbao-openbao-test
+                name: unit-test
         status:
           atProvider:
             manifest:
               apiVersion: v1
-              kind: Secret
+              kind: Namespace
               metadata:
-                name: openbao-test-init-output
-                namespace: vshn-openbao-openbao-test
-              data:
-                VAULT_ADDR: aHR0cHM6Ly9vcGVuYmFvLXRlc3Q6ODIwMA==
-                VAULT_TOKEN: cy5zb21ldG9rZW4=
+                name: unit-test
+                labels:
+                  appuio.io/organization: vshn
+    openbao-test-ns:
+      resource:
+        apiVersion: kubernetes.crossplane.io/v1alpha2
+        kind: Object
+        metadata:
+          name: vshn-openbao-openbao-test
+        spec:
+          forProvider:
+            manifest:
+              apiVersion: v1
+              kind: Namespace
+              metadata:
+                name: vshn-openbao-openbao-test
+        status:
+          atProvider:
+            manifest:
+              apiVersion: v1
+              kind: Namespace
+              metadata:
+                name: vshn-openbao-openbao-test
