+++ test/functions/vshnopenbao/deploy/04_status_initialized.yaml
@@ -35,29 +35,5 @@ observed:
             authClients: true
       status:
         instanceNamespace: vshn-openbao-openbao-test
-  resources:
-    openbao-test-init-output-observer:
-      resource:
-        apiVersion: kubernetes.crossplane.io/v1alpha2
-        kind: Object
-        metadata:
-          name: openbao-test-init-output
-        spec:
-          forProvider:
-            manifest:
-              apiVersion: v1
-              kind: Secret
-              metadata:
-                name: openbao-test-init-output
-                namespace: vshn-openbao-openbao-test
-        status:
-          atProvider:
-            manifest:
-              apiVersion: v1
-              kind: Secret
-              metadata:
-                name: openbao-test-init-output
-                namespace: vshn-openbao-openbao-test
-              data:
-                VAULT_ADDR: aHR0cHM6Ly9vcGVuYmFvLXRlc3Q6ODIwMA==
-                VAULT_TOKEN: cy5zb21ldG9rZW4=
+        initialized: true
+  resources: {}
